Fixed some calculator output type comments. Moved enums into normal TS file. Added missing ELM attribute.
